requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-filter==22.1
django-oauth2-provider==0.2.6.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
pytz==2022.6
shortuuid==1.0.9
sqlparse==0.4.3
zipp==3.10.0
